FindProductsByFilterUseCase in one file.
